requirements.txt
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.36
pandas-ta==0.3.14b
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
alpaca-py==0.20.0
python-telegram-bot==20.7
